implementar paginación en nuestrosAnimales.html

+++ src/main/java/com/grupo3/pawHome/repositories/FacturaRepository.java
@@ -3,5 +3,5 @@ package com.grupo3.pawHome.repositories;
 import com.grupo3.pawHome.entities.Factura;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FacturaRepository extends JpaRepository<Factura, Long> {
+public interface FacturaRepository extends JpaRepository<Factura, Integer> {
 }
